// src/models/SaasCsp2OfficeTenant.ts
/* WARNING: This file is auto-generated . Do not edit manually. */

import { SaasCsp2ServiceStateEnum } from './SaasCsp2ServiceStateEnum';

export type SaasCsp2OfficeTenant = {
  address?: string;
  city?: string;
  creationDate?: Date;
  displayName?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  mpnId?: string | null;
  mpnIssueFlag?: string | null;
  phone?: string;
  serviceName?: string;
  status?: SaasCsp2ServiceStateEnum;
  zipCode?: string;
};